# api/openapi.yaml
openapi: 3.0.0
paths:
  /api/v1/auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponseDto'
        '409':
          description: EMAIL_ALREADY_REGISTERED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      summary: Crea una cuenta y devuelve la sesion iniciada
      tags:
        - auth
  /api/v1/auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponseDto'
        '401':
          description: INVALID_CREDENTIALS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      summary: Inicia sesion en un dispositivo
      tags:
        - auth
  /api/v1/auth/refresh:
    post:
      description: >-
        Invalida el token usado y emite uno nuevo con 60 dias desde ahora. Si llega un token ya
        rotado responde TOKEN_REUSE_DETECTED y revoca todas las sesiones del dispositivo.
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairDto'
        '401':
          description: INVALID_REFRESH_TOKEN o TOKEN_REUSE_DETECTED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      summary: Rota el refresh token
      tags:
        - auth
  /api/v1/auth/logout:
    post:
      description: 'Idempotente: cerrar una sesion ya cerrada responde ok igual.'
      operationId: AuthController_logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutDto'
      responses:
        '200':
          description: ''
      summary: Cierra la sesion de este dispositivo
      tags:
        - auth
  /api/v1/auth/sessions:
    get:
      operationId: AuthController_sessions
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionDto'
      security:
        - access-token: []
      summary: Lista los dispositivos con sesion activa
      tags:
        - auth
  /api/v1/auth/sessions/{id}:
    delete:
      operationId: AuthController_revokeSession
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Cierra la sesion de otro dispositivo
      tags:
        - auth
  /api/v1/auth/forgot-password:
    post:
      description: >-
        Responde siempre lo mismo, exista el email o no: si distinguiera, seria un comprobador
        gratuito de quien tiene cuenta.
      operationId: AuthController_forgotPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordDto'
      responses:
        '200':
          description: ''
      summary: Pide un enlace de recuperacion
      tags:
        - auth
  /api/v1/auth/reset-password:
    post:
      description: 'Cierra TODAS las sesiones del usuario: si alguien habia entrado, queda fuera.'
      operationId: AuthController_resetPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '200':
          description: ''
      summary: Cambia la contrasena con el token del correo
      tags:
        - auth
  /api/v1/auth/me:
    get:
      operationId: AuthController_me
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserDto'
      security:
        - access-token: []
      summary: Datos del usuario autenticado
      tags:
        - auth
...
info:
  title: PaceUp API
  description: |-
    Backend de PaceUp: maratones, inscripciones, planes de entrenamiento y tracking GPS.

    **Unidades base:** distancias en metros, duraciones en segundos, dinero en centavos
    (entero) con `currency`, fechas en ISO-8601 UTC. El formateo (km, min/km, `Bs`) es del cliente.

    **Sobre de respuesta:** todo exito devuelve `{ data, meta }` y todo error `{ error, meta }`.
    Mapea los errores por `error.code`, nunca por `error.message`.
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local
  - url: https://api.paceup.example
    description: VPS
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
    ingest-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: El `ingestToken` que devuelve POST /workouts/sessions
  schemas:
    RegisterDto:
      type: object
      properties:
        deviceId:
          type: string
          example: 9f1c8d2e-4b7a-4c11-9f3e-2a1b5c6d7e8f
        deviceName:
          type: string
          example: Pixel 8 de Alvaro
        platform:
          type: string
          example: android
        email:
          type: string
          example: runner@test.com
        password:
          type: string
          example: Test1234!
          minLength: 8
        name:
          type: string
          example: Alvaro Quispe
      required:
        - deviceId
        - email
        - password
        - name
    AuthUserDto:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum:
            - runner
            - admin
        onboardingSeenAt:
          type: object
          nullable: true
          description: Null = el usuario aun no vio el onboarding. Sobrevive a reinstalaciones.
      required:
        - id
        - email
        - name
        - role
        - onboardingSeenAt
    AuthSessionResponseDto:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
          description: Cadena opaca. Guardala en flutter_secure_storage.
        expiresIn:
          type: number
          example: 900
          description: Segundos de vida del access token
        user:
          $ref: '#/components/schemas/AuthUserDto'
      required:
        - accessToken
        - refreshToken
        - expiresIn
        - user
    ApiErrorDto:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
          description: Codigo estable; mapea por este campo, no por el mensaje
        message:
          type: string
          example: El campo email no es valido
        details:
          example: []
          type: array
          items:
            type: object
      required:
        - code
        - message
        - details
    ResponseMetaDto:
      type: object
      properties:
        requestId:
          type: string
          example: c1f3a0d2-9b7e-4a11-8f2c-5d6e7a8b9c0d
        timestamp:
          type: string
          example: '2026-08-18T14:03:22.118Z'
          description: Hora del servidor, ISO-8601 UTC
        nextCursor:
          type: object
          nullable: true
          description: Cursor de la pagina siguiente
      required:
        - requestId
        - timestamp
    ErrorResponseDto:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ApiErrorDto'
        meta:
          $ref: '#/components/schemas/ResponseMetaDto'
      required:
        - error
        - meta
    LoginDto:
      type: object
      properties:
        deviceId:
          type: string
          example: 9f1c8d2e-4b7a-4c11-9f3e-2a1b5c6d7e8f
        deviceName:
          type: string
          example: Pixel 8 de Alvaro
        platform:
          type: string
          example: android
        email:
          type: string
          example: runner@test.com
        password:
          type: string
          example: Test1234!
      required:
        - deviceId
        - email
        - password
    RefreshDto:
      type: object
      properties:
        deviceId:
          type: string
          example: 9f1c8d2e-4b7a-4c11-9f3e-2a1b5c6d7e8f
        deviceName:
          type: string
          example: Pixel 8 de Alvaro
        platform:
          type: string
          example: android
        refreshToken:
          type: string
          description: El refresh token entregado en el login o en el refresh anterior
      required:
        - deviceId
        - refreshToken
    TokenPairDto:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
          description: Cadena opaca. Guardala en flutter_secure_storage.
        expiresIn:
          type: number
          example: 900
          description: Segundos de vida del access token
      required:
        - accessToken
        - refreshToken
        - expiresIn
    LogoutDto:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    SessionDto:
      type: object
      properties:
        id:
          type: string
        deviceId:
          type: string
        deviceName:
          type: object
          nullable: true
        platform:
          type: object
          nullable: true
        lastUsedAt:
          type: string
        expiresAt:
          type: string
        current:
          type: boolean
          description: True si es la sesion desde la que estas llamando
      required:
        - id
        - deviceId
        - deviceName
        - platform
        - lastUsedAt
        - expiresAt
        - current
    ForgotPasswordDto:
      type: object
      properties:
        email:
          type: string
          example: runner@test.com
      required:
        - email
    ResetPasswordDto:
      type: object
      properties:
        token:
          type: string
          description: El token que viaja en el enlace del correo
        password:
          type: string
          example: NuevaClave123
          minLength: 8
      required:
        - token
        - password
...